Inv logic in inventory.{c,h}, create craft.h.
