+++ src/inventory.c
@@ -1,3 +1,13 @@
+#include <GL/glew.h>
+#include <GLFW/glfw3.h>
+#include <stdio.h>
+#include <string.h>
+#include "config.h"
+#include "craft.h"
+#include "inventory.h"
+
+Inventory inventory;
+
 // From https://github.com/CouleeApps/Craft/tree/mining_crafting
 void make_inventory(float *data, float x, float y, float n, float m, int s) {
     float *d = data;
@@ -20,3 +30,111 @@ void make_inventory(float *data, float x, float y, float n, float m, int s) {
     *(d++) = x - n; *(d++) = y + m;
     *(d++) = du + 0; *(d++) = b - p;
 }
+
+// From https://github.com/CouleeApps/Craft/tree/mining_crafting
+GLuint gen_inventory_buffers(float x, float y, float n, int sel) {
+    int length = INVENTORY_SLOTS;
+    GLfloat *data = malloc_faces(4, length);
+    x -= n * (length - 1) / 2;
+    for (int i = 0; i < length; i ++) {
+        make_inventory(data + i * 24, x, y, n / 2, n / 2, sel == i ? 1 : 0);
+        x += n;
+    }
+    return gen_faces(4, length, data);
+}
+
+// From https://github.com/CouleeApps/Craft/tree/mining_crafting
+void draw_inventory(Attrib *attrib, GLuint buffer, int length) {
+    glEnable(GL_BLEND);
+    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+    draw_triangles_2d(attrib, buffer, length * 6);
+    glDisable(GL_BLEND);
+}
+
+// Modified version from https://github.com/CouleeApps/Craft/tree/mining_crafting
+void render_inventory_item(Attrib *attrib, Item item, float xpos, float ypos,
+        int width, int height, int sel) {
+    glUseProgram(attrib->program);
+    glUniform3f(attrib->camera, 0, 0, 5);
+    glUniform1i(attrib->sampler, 0); // GL_TEXTURE0
+    glUniform1f(attrib->timer, PI*2);
+    float matrix[16];
+    GLuint buffer;
+    set_matrix_item_offs(matrix, width, height, 0.65, xpos, ypos, sel);
+    if (is_plant(item.id)) {
+        glDeleteBuffers(1, &buffer);
+        buffer = gen_plant_buffer(0, 0, 0, 0.5, item.id);
+    } else {
+        buffer = gen_cube_buffer(0, 0, 0, 0.5, item.id);
+    }
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    if (is_plant(item.id)) {
+        draw_plant(attrib, buffer);
+    } else {
+        draw_cube(attrib, buffer);
+    }
+    del_buffer(buffer);
+}
+
+void render_inventory_items(Attrib *attrib, float pos_x, int width, int height) {
+    for (int item = 0; item < INVENTORY_SLOTS; item ++) {
+        Item block = inventory.items[item];
+        if (block.id == 0 || block.num == 0) continue;
+
+        float slotoff = -1 *  ((float)item - (float)(INVENTORY_SLOTS - 1) / 2);
+        float xpos = slotoff * ((0.125*1024)/width);
+        float ypos = 1 - ((0.17*768)/height);
+        int sel = inventory.selected == item ? 1 : 0;
+        render_inventory_item(attrib, block, xpos, ypos, width, height, sel);
+    }
+}
+
+// Modified version from https://github.com/CouleeApps/Craft/tree/mining_crafting
+void render_inventory_bar(Attrib *attrib, float x, int sel,
+        int width, int height) {
+    float matrix[16];
+    set_matrix_2d(matrix, width, height);
+    glClear(GL_DEPTH_BUFFER_BIT);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform1i(attrib->sampler, 4); // GL_TEXTURE4
+    GLuint inv_buffer = gen_inventory_buffers(x, 64, 64, sel);
+    draw_inventory(attrib, inv_buffer, INVENTORY_SLOTS);
+    del_buffer(inv_buffer);
+}
+
+// Modified version from https://github.com/CouleeApps/Craft/tree/mining_crafting
+void render_inventory_text(Attrib *attrib, Item item, float x, float y,
+        int width, int height) {
+    float matrix[16];
+    set_matrix_2d(matrix, width, height);
+    glUseProgram(attrib->program);
+    glUniformMatrix4fv(attrib->matrix, 1, GL_FALSE, matrix);
+    glUniform1i(attrib->sampler, 1); // GL_TEXTURE1
+    char text_buffer[16];
+    float ts = 12;
+    snprintf(text_buffer, 16, "%02d", item.num);
+    x += ts * strlen(text_buffer);
+    print(attrib, 1, x, y, ts, text_buffer);
+}
+
+void render_inventory_texts(Attrib *attrib, float x, int width, int height) {
+    for (int item = 0; item < INVENTORY_SLOTS; item ++) {
+        Item block = inventory.items[item];
+        if (block.id == 0 || block.num <= 0) continue;
+
+        float tx = 12 + x + 64 * (item - ((float)INVENTORY_SLOTS / 2.));
+        float ty = 20;
+        render_inventory_text(attrib, block, tx, ty, width, height);
+    }
+}
+
+void render_inventory(Attrib *window_attrib, Attrib *block_attrib, Attrib *text_attrib,
+        float pos_x, int sel, int width, int height) {
+    render_inventory_bar(window_attrib, pos_x, sel, width, height);
+    glClear(GL_DEPTH_BUFFER_BIT);
+    render_inventory_items(block_attrib, pos_x, width, height);
+    glClear(GL_DEPTH_BUFFER_BIT);
+    render_inventory_texts(text_attrib, pos_x, width, height);
+}
+
